implementação de erros costumizados nos controllers

+++ users/views.py
@@ -2,12 +2,11 @@ from typing import Any, Dict, cast
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from django.utils import timezone
-from users.serializers import UserSerializer, PasswordUpdateSerializer, PasswordRecoverySerializer, ConfirmAccountSerializer
+from users.serializers import UserSerializer, PasswordUpdateSerializer, PasswordRecoverySerializer, ConfirmAccountSerializer, UserUpadateDataSerializer
+from users.helpers.errors.error import BadRequestError, ConflictError, NotFoundError, UnauthorizedError, DatabaseError,AppError
 
 from users.services.user_service import UserService
 from users.infra.userRepository import UserRepository
-
 
 class UserController(APIView):
     def __init__(self, **kwargs: Any) -> None:
@@ -20,7 +19,7 @@ class UserController(APIView):
 
             # valida e retorna 400 com erros do serializer se inválido
             if not serializer.is_valid():
-                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                raise BadRequestError(safe_message="Invalid data", status_code=400, extra=serializer.errors, code="bad_request")
 
             # informa ao type checker que validated_data é dict
             validated_data = cast(Dict[str, Any], serializer.validated_data)
@@ -36,9 +35,12 @@ class UserController(APIView):
 
         except ValueError as e:
             # Erro de regra de negócio (ex.: email duplicado)
-            return Response({"error": str(e)}, status=status.HTTP_409_CONFLICT)
+            raise ConflictError(safe_message="E-mail or user alrady exists", status_code=409, code="conflict")
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": "Internal server error"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
         
     def get(self, request):
         try:
@@ -53,14 +55,22 @@ class UserController(APIView):
             )
 
         except Exception as e:
-            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
 
     def put(self, request, id: str):
         try:
-            serializer = UserSerializer(data=request.data, partial=True)
+            if id == None:
+                raise UnauthorizedError(safe_message="Not authorized to proceed with the operation", status_code=401, code="unauthorized")
+            
+                user_id = self.service.get_user_by_id(id)
+
+                if user_id is None:
+                    raise NotFoundError(safe_message="User not found", status_code=404, code="not_found")
+            
+            serializer = UserUpadateDataSerializer(data=request.data, partial=True)
 
             if not serializer.is_valid():
-                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                raise BadRequestError(safe_message="invalid data", status_code=400, code="bad_request", extra=serializer.errors)
 
             validated_data = cast(Dict[str, Any], serializer.validated_data)
 
@@ -71,39 +81,53 @@ class UserController(APIView):
                 validated_data.get("email"),
             )
 
-            if user is None:
-                return Response({"error": "Usuário não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+            if not user:
+                raise NotFoundError(safe_message="User not found", status_code=404, code="not_found")
 
             return Response(UserSerializer(user.__dict__).data, status=status.HTTP_200_OK)
 
         except ValueError as e:
-            return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+            raise BadRequestError(safe_message="Something unexpected happened", status_code=400, code="bad_request")
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": "Internal server error"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
         
     def delete(self, request, id: str):
         try:
-            self.service.delete_user_by_id(id)
+            if not id:
+                raise UnauthorizedError(safe_message="Not authorized to proceed with the operation", status_code=401, code="unauthorized")
+
+            userId = self.service.delete_user_by_id(id)
+
+            if not userId:
+                raise NotFoundError(safe_message="user not found", status_code=404, code="not_found")
 
             return Response(status=status.HTTP_204_NO_CONTENT)
 
         except ValueError as e:
-            return Response({"error": str(e)}, status=status.HTTP_404_NOT_FOUND)
+            raise BadRequestError(safe_message="Something unexpected happened", status_code=400, code="bad_request")
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
 
-
-
-class PasswordController(APIView):
+class SpecificUserController(APIView):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(**kwargs)
         self.service = UserService(UserRepository())
 
     def get(self, request, id: str): 
         try: 
+            if not id:
+                raise UnauthorizedError(safe_message="Not authorized to proceed with the operation", status_code=401, code="unauthorized")
+            
             user = self.service.get_user_by_id(id) 
-            if user is None:
-                return Response({"error": "Usuário não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+
+            if not user :
+                raise NotFoundError(safe_message="User not found", status_code=404, code="not_found")
             
             return Response({
             "id": str(user.id),
@@ -113,16 +137,24 @@ class PasswordController(APIView):
             "created_at": user.created_at.isoformat(),
             "updated_at": user.updated_at.isoformat(),
         }, status=status.HTTP_200_OK)
-
+        
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
+
+class PasswordController(APIView):
+    def __init__(self, **kwargs: Any) -> None:
+        super().__init__(**kwargs)
+        self.service = UserService(UserRepository())
 
     def post(self, request):
         try:
             serializer = PasswordRecoverySerializer(data=request.data)
 
             if not serializer.is_valid():
-                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                raise BadRequestError(safe_message="Invalid data", status_code=400, code="bad_request", extra=serializer.errors)
 
             validated_data = cast(Dict[str, Any], serializer.validated_data)
 
@@ -130,22 +162,27 @@ class PasswordController(APIView):
                 validated_data["email"],
                 validated_data["password"],
             )
+
             if user is None:
-                return Response({"error": "Usuário não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+                raise NotFoundError(safe_message="user not found", status_code=404, code="not_found")
 
             return Response(UserSerializer(user.__dict__).data, status=status.HTTP_200_OK)
 
-        except ValueError as e:
-            return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": "Internal server error"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
 
     def put(self, request, id:str):
         try:
+            if not id:
+                raise UnauthorizedError(safe_message="Not authorized to proceed with the operation", status_code=401, code="unauthorized")
+            
             serializer = PasswordUpdateSerializer(data=request.data)
 
             if not serializer.is_valid():
-                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+                raise BadRequestError(safe_message="invalid data",status_code=400, code="bad_request", extra=serializer.errors)
 
             validated_data = cast(Dict[str, Any], serializer.validated_data)
 
@@ -154,16 +191,17 @@ class PasswordController(APIView):
                 validated_data["old_password"],
                 validated_data["new_password"],
             )
+
             if user is None:
-                return Response({"error": "Usuário não encontrado"}, status=status.HTTP_404_NOT_FOUND)
+                raise NotFoundError(safe_message="user not found", status_code=404, code="not_found")
 
             return Response(UserSerializer(user.__dict__).data, status=status.HTTP_200_OK)
 
-        except ValueError as e:
-            # ex: senha antiga incorreta
-            return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": "Internal server error"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
 
 class ConfirmAccountView(APIView):
 
@@ -175,7 +213,7 @@ class ConfirmAccountView(APIView):
         serializer = ConfirmAccountSerializer(data=request.data)
 
         if not serializer.is_valid():
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            raise BadRequestError(safe_message="invalid data", status_code=400, code="bad_request", extra=serializer.errors)
         
         validated_data = cast(Dict[str, Any], serializer.validated_data)
 
@@ -186,9 +224,12 @@ class ConfirmAccountView(APIView):
             user = self.service.confirm_user(email, code)
 
             if user is None:
-                return Response({"error": "Usuário não encontrado 🚫"}, status=status.HTTP_404_NOT_FOUND)
+                raise NotFoundError(safe_message="user not found", status_code=404, code="not_found")
 
-            return Response({"message": "Conta confirmada com sucesso 🎉"}, status=status.HTTP_200_OK)
+            return Response({"message": "Account successfully confirmed 🎉"}, status=status.HTTP_200_OK)
         
+        except AppError:
+            # deixa o middleware global tratar os AppError customizados
+            raise
         except Exception as e:
-            return Response({"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            raise DatabaseError(safe_message="Something went wrong on our side, please try again later, and if it persists, contact us", status_code=500, code="internal_server_error")
